requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
telethon==1.33.1
python-dotenv==1.0.0
cryptg==0.4.0
